Add simple sparse matrix implementation.
A simple sparse matrix implementation to be used with creating a
lookup table where keys are pairs of ints and values are ints.
This is intended to be used with mapping nodes_at_link <-> links.
That is, given two nodes, what is the id of the link connecting
them, if there is one.